+++ src/main/java/com/zeylin/eventsourcingwithcqrs/aggregates/ShowtimeAggregate.java
@@ -4,16 +4,24 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import javax.persistence.Entity;
+import javax.persistence.Id;
+import javax.validation.constraints.NotNull;
 import java.time.LocalDateTime;
 
+@Entity
 @Getter
 @Setter
 @NoArgsConstructor
 public class ShowtimeAggregate {
 
+    @Id
     private String id;
+    @NotNull
     private Integer movieId;
+    @NotNull
     private LocalDateTime time;
+    @NotNull
     private Integer auditoriumId;
 
 }