+++ src/main/java/com/zeylin/eventsourcingwithcqrs/aggregates/ShowtimeAggregateEventHandler.java
@@ -2,6 +2,8 @@ package com.zeylin.eventsourcingwithcqrs.aggregates;
 
 import com.zeylin.eventsourcingwithcqrs.events.ShowtimeCreatedEvent;
 import com.zeylin.eventsourcingwithcqrs.events.ShowtimeUpdatedEvent;
+import com.zeylin.eventsourcingwithcqrs.exceptions.NotFoundException;
+import com.zeylin.eventsourcingwithcqrs.repositories.ShowtimeAggregateRepository;
 import com.zeylin.eventsourcingwithcqrs.utils.annotations.EventHandler;
 import com.zeylin.eventsourcingwithcqrs.utils.EventType;
 import org.slf4j.Logger;
@@ -13,28 +15,35 @@ public class ShowtimeAggregateEventHandler {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(ShowtimeAggregateEventHandler.class);
 
-    private ShowtimeAggregate showtimeAggregate;
+    private final ShowtimeAggregateRepository showtimeAggregateRepository;
 
-    public ShowtimeAggregateEventHandler() {
+    public ShowtimeAggregateEventHandler(ShowtimeAggregateRepository showtimeAggregateRepository) {
+        this.showtimeAggregateRepository = showtimeAggregateRepository;
     }
 
     @EventHandler(type = EventType.SHOWTIME_CREATE)
     public void on(ShowtimeCreatedEvent event) {
         LOGGER.info("Received event of type {}", EventType.SHOWTIME_CREATE);
 
-        showtimeAggregate = new ShowtimeAggregate();
+        ShowtimeAggregate showtimeAggregate = new ShowtimeAggregate();
         showtimeAggregate.setId(event.getId());
         showtimeAggregate.setMovieId(event.getMovieId());
         showtimeAggregate.setAuditoriumId(event.getAuditoriumId());
         showtimeAggregate.setTime(event.getTime());
+
+        showtimeAggregateRepository.save(showtimeAggregate);
     }
 
     @EventHandler(type = EventType.SHOWTIME_UPDATE)
-    public void on(ShowtimeUpdatedEvent event) {
+    public void on(ShowtimeUpdatedEvent event) throws NotFoundException {
         LOGGER.info("Received event of type {}", EventType.SHOWTIME_UPDATE);
 
+        ShowtimeAggregate showtimeAggregate = showtimeAggregateRepository.findById(event.getId())
+                .orElseThrow(() -> new NotFoundException("Showtime not found."));
         showtimeAggregate.setId(event.getId());
         showtimeAggregate.setTime(event.getTime());
+
+        showtimeAggregateRepository.save(showtimeAggregate);
     }
 
 }