+++ src/main/java/com/zeylin/eventsourcingwithcqrs/exceptions/NotFoundException.java
@@ -1,0 +1,9 @@
+package com.zeylin.eventsourcingwithcqrs.exceptions;
+
+public class NotFoundException extends Exception {
+
+    public NotFoundException(String s) {
+        super(s);
+    }
+
+}