+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.1.0-SNAPSHOT] - 2020-03-24
+### Added
+- Persisting showtime aggregate on events
+- Not Found Exception
+
+### Changed
+- ShowtimeAggregate -> entity
+
 ## [1.0.0-SNAPSHOT] - 2020-03-23
 ### Added
 - Aggregate for showtimes